+++ llama-cpp-4/src/context/kv_cache.rs
@@ -92,21 +92,10 @@ impl LlamaContext<'_> {
         let p1 = p1
             .map_or(Ok(-1), i32::try_from)
             .map_err(KvCacheConversionError::P1TooLarge)?;
-        #[cfg(feature = "mtp")]
-        {
-            Ok(
-                unsafe {
-                    llama_cpp_sys_4::llama_context_seq_rm(self.context.as_ptr(), src, p0, p1)
-                },
-            )
-        }
-        #[cfg(not(feature = "mtp"))]
-        {
-            Ok(unsafe {
-                let mem = llama_cpp_sys_4::llama_get_memory(self.context.as_ptr());
-                llama_cpp_sys_4::llama_memory_seq_rm(mem, src, p0, p1)
-            })
-        }
+        Ok(unsafe {
+            let mem = llama_cpp_sys_4::llama_get_memory(self.context.as_ptr());
+            llama_cpp_sys_4::llama_memory_seq_rm(mem, src, p0, p1)
+        })
     }
 
     /// Clear the KV cache